Move activation-only config to PARQ prototype
