+++ test/prototype/test_parq.py
@@ -26,12 +26,14 @@ from torchao.prototype.parq.quant import (
     UnifQuantizer,
     UnifTorchaoQuantizer,
 )
-from torchao.prototype.parq.quant.quant_api import StretchedIntxWeightOnlyConfig
+from torchao.prototype.parq.quant.quant_api import (
+    Int8DynamicActivationOnlyConfig,
+    StretchedIntxWeightOnlyConfig,
+)
 from torchao.prototype.parq.quant.uniform_torchao import _BIT_WIDTH_TO_DTYPE
 from torchao.quantization.granularity import PerGroup
 from torchao.quantization.qat import QATConfig
 from torchao.quantization.quant_api import (
-    Int8DynActOnlyConfig,
     Int8DynamicActivationIntxWeightConfig,
     IntxWeightOnlyConfig,
     _is_linear,
@@ -392,7 +394,7 @@ class TestInt8DynamicActivationTorchaoQuantizer(common_utils.TestCase):
         optimizer.step()
 
         # apply torchao quantized activations on top
-        qat_config = QATConfig(Int8DynActOnlyConfig(), step="prepare")
+        qat_config = QATConfig(Int8DynamicActivationOnlyConfig(), step="prepare")
         filter_fn = optimizer.get_filter_fn(model)
         quantize_(model, qat_config, filter_fn=filter_fn)
         out = model(x)
